refactor: nouvelle archi complète pour la partie campus
