+++ src/features/Campus/Dashboard/components/FreeRoomSection.tsx
@@ -10,9 +10,10 @@ import { CampusDataManager as DataManager } from '../../services/CampusDataManag
 import { getDistanceInKm, BuildingInfo } from '../../services/FreeRoomService';
 import { useFavorites } from '../../hooks/useFavorites';
 
-const defaultImage = require('../../../../../assets/images/default_resto.png');
 const { width } = Dimensions.get('window');
 const CARD_WIDTH = width * 0.85;
+
+import { FreeRoomSectionCard } from './FreeRoomSectionCard';
 
 export function FreeRoomSection({ navigation, userLat, userLon }: { navigation: import('@react-navigation/native').NavigationProp<Record<string, unknown>>, userLat?: number, userLon?: number }) {
     const { themeName } = useContext(AppContext);
@@ -68,74 +69,13 @@ export function FreeRoomSection({ navigation, userLat, userLon }: { navigation: 
     }, [buildings, favBuildings]);
 
     const renderCard = ({ item }: { item: BuildingInfo }) => {
-        const imageSource = item.imageUrl ? { uri: item.imageUrl } : defaultImage;
-        const totalRooms = item.rooms ? item.rooms.length : 0;
-        
-        let hoursText = Translator.get('UNKNOWN') || 'Non communiqué';
-        if (item.schedule) {
-            const currentDay = new Date().getDay() || 7; // 1-7
-            const daySchedule = item.schedule[String(currentDay)];
-            if (daySchedule) {
-                hoursText = `${daySchedule.open} - ${daySchedule.close}`;
-            } else {
-                hoursText = Translator.get('BU_CLOSED') || 'Fermé';
-            }
-        }
-        
         return (
-            <Reanimated.View 
-                entering={FadeIn}
-                layout={LinearTransition.springify()}
-            >
-                <TouchableOpacity 
-                    activeOpacity={0.9}
-                    onPress={() => navigation.navigate('FreeRoomDetails', { building: item })}
-                    style={{
-                        width: CARD_WIDTH,
-                        backgroundColor: theme.cardBackground,
-                        borderRadius: tokens.radius.xl, 
-                        marginRight: tokens.space.md,
-                        ...tokens.shadow.md, 
-                        overflow: 'hidden', 
-                    }}
-                >
-                    <Image source={imageSource} style={{ width: '100%', height: 160, resizeMode: 'cover', backgroundColor: theme.greyBackground }} />
-
-                    <View style={{ padding: tokens.space.md }}>
-                        <View style={{ flexDirection: 'row', alignItems: 'center', marginBottom: tokens.space.xs }}>
-                            <Text style={{ fontSize: tokens.fontSize.lg, fontWeight: tokens.fontWeight.bold, color: theme.font, flexShrink: 1 }} numberOfLines={1}>
-                                {item.name}
-                            </Text>
-                            <TouchableOpacity onPress={() => toggleFavBuilding(item.id)} hitSlop={{ top: 15, bottom: 15, left: 10, right: 15 }} style={{ marginLeft: 6 }}>
-                                <MaterialCommunityIcons name={favBuildings.includes(item.id) ? "star" : "star-outline"} size={22} color={favBuildings.includes(item.id) ? theme.primary : theme.fontSecondary} />
-                            </TouchableOpacity>
-                        </View>
-                        
-                        <View style={{ flexDirection: 'row', alignItems: 'center', marginBottom: tokens.space.xs }}>
-                            <MaterialIcons name="location-on" size={16} color={theme.fontSecondary} />
-                            <Text style={{ fontSize: tokens.fontSize.sm, color: theme.fontSecondary, marginLeft: 4, flex: 1 }} numberOfLines={1}>
-                                {item.campus || 'Talence'}
-                            </Text>
-
-                            {item.distance !== undefined && (
-                                <View style={{ flexDirection: 'row', alignItems: 'center', backgroundColor: `${theme.primary}15`, paddingHorizontal: tokens.space.sm, paddingVertical: 4, borderRadius: tokens.radius.md }}>
-                                    <MaterialCommunityIcons name="walk" size={14} color={theme.primary} />
-                                    <Text style={{ fontSize: tokens.fontSize.sm, fontWeight: tokens.fontWeight.bold, color: theme.primary, marginLeft: 4 }}>
-                                        {item.distance < 1 ? `${Math.round(item.distance * 1000)} m` : `${item.distance.toFixed(1)} km`}
-                                    </Text>
-                                </View>
-                            )}
-                        </View>
-                        
-                        <View style={{ flexDirection: 'row', alignItems: 'center' }}>
-                            <MaterialCommunityIcons name="clock-outline" size={16} color={theme.fontSecondary} />
-                            <Text style={{ fontSize: tokens.fontSize.sm, color: theme.fontSecondary, marginLeft: 4, flex: 1 }}>
-                                {hoursText} • {totalRooms} {Translator.get('ROOMS' as Parameters<typeof Translator.get>[0]) || 'Salles'}
-                            </Text>
-                        </View>
-                    </View>
-                </TouchableOpacity>
-            </Reanimated.View>
+            <FreeRoomSectionCard 
+                item={item} 
+                navigation={navigation} 
+                isFavorite={favBuildings.includes(item.id)} 
+                onToggleFavorite={toggleFavBuilding} 
+            />
         );
     };
