+++ src/features/Campus/Dashboard/components/LibrarySection.tsx
@@ -10,9 +10,10 @@ import LibraryService, { LibraryInfo, AffluencesData } from '../../services/Libr
 import { useFavorites } from '../../hooks/useFavorites';
 import { useSavedFilter } from '../../hooks/useSavedFilter';
 
-const defaultBuImage = require('../../../../../assets/images/default_resto.png');
 const { width } = Dimensions.get('window');
 const CARD_WIDTH = width * 0.85;
+
+import { LibrarySectionCard } from './LibrarySectionCard';
 
 export function LibrarySection({ navigation, userLat, userLon }: { navigation: import('@react-navigation/native').NavigationProp<Record<string, unknown>>, userLat?: number, userLon?: number }) {
     const { themeName } = useContext(AppContext);
@@ -73,89 +74,14 @@ export function LibrarySection({ navigation, userLat, userLon }: { navigation: i
     }, [libraries, favBu, libraryFilter, affluences]);
 
     const renderCard = ({ item }: { item: LibraryInfo }) => {
-        const affluenceData = affluences[item.id];
-        const rate = affluenceData?.occupancyRate ?? null;
-        const isOpen = affluenceData?.isOpen ?? true;
-
-        let statusColor = '#f44336';
-        if (isOpen) {
-            if (rate === null || rate < 50) statusColor = '#4caf50';
-            else if (rate < 80) statusColor = '#ff9800';
-            else statusColor = '#ff4436';
-        }
-
-        let statusText = isOpen ? (Translator.get('BU_OPEN') || 'Ouvert') : (Translator.get('BU_CLOSED') || 'Fermé');
-        if (!isOpen && affluenceData?.openingText) {
-            statusText = `${statusText} - ${affluenceData.openingText}`;
-        }
-
-        const imageSource = item.imageUrl ? { uri: item.imageUrl } : defaultBuImage;
-
         return (
-            <Reanimated.View
-                entering={FadeIn}
-                layout={LinearTransition.springify()}
-            >
-                <TouchableOpacity
-                    activeOpacity={0.9}
-                    onPress={() => navigation.navigate('LibraryDetails', { library: item, affluence: affluenceData })}
-                    style={{
-                        width: CARD_WIDTH,
-                        backgroundColor: theme.cardBackground,
-                        borderRadius: tokens.radius.xl,
-                        marginRight: tokens.space.md,
-                        ...tokens.shadow.md,
-                        overflow: 'hidden',
-                    }}
-                >
-                    <Image source={imageSource} style={{ width: '100%', height: 160, resizeMode: 'cover', backgroundColor: theme.greyBackground }} />
-
-                    <View style={{ padding: tokens.space.md }}>
-                        <View style={{ flexDirection: 'row', alignItems: 'center', marginBottom: tokens.space.xs }}>
-                            <Text style={{ fontSize: tokens.fontSize.lg, fontWeight: tokens.fontWeight.bold, color: theme.font, flexShrink: 1 }} numberOfLines={1}>
-                                {item.name}
-                            </Text>
-                            <TouchableOpacity onPress={() => toggleFavBu(item.id)} hitSlop={{ top: 15, bottom: 15, left: 10, right: 15 }} style={{ marginLeft: 6 }}>
-                                <MaterialCommunityIcons name={favBu.includes(item.id) ? "star" : "star-outline"} size={22} color={favBu.includes(item.id) ? theme.primary : theme.fontSecondary} />
-                            </TouchableOpacity>
-                        </View>
-
-                        <View style={{ flexDirection: 'row', alignItems: 'center', marginBottom: tokens.space.xs }}>
-                            <MaterialIcons name="location-on" size={16} color={theme.fontSecondary} />
-                            <Text style={{ fontSize: tokens.fontSize.sm, color: theme.fontSecondary, marginLeft: 4, flex: 1 }} numberOfLines={1}>
-                                {item.campus}
-                            </Text>
-
-                            {item.distance !== undefined && (
-                                <View style={{ flexDirection: 'row', alignItems: 'center', backgroundColor: `${theme.primary}15`, paddingHorizontal: tokens.space.sm, paddingVertical: 4, borderRadius: tokens.radius.md }}>
-                                    <MaterialIcons name="directions-walk" size={14} color={theme.primary} />
-                                    <Text style={{ fontSize: tokens.fontSize.sm, fontWeight: tokens.fontWeight.bold, color: theme.primary, marginLeft: 4 }}>
-                                        {item.distance < 1 ? `${Math.round(item.distance * 1000)} m` : `${item.distance.toFixed(1)} km`}
-                                    </Text>
-                                </View>
-                            )}
-                        </View>
-
-                        <View style={{ flexDirection: 'row', alignItems: 'center' }}>
-                            <MaterialCommunityIcons name={isOpen ? 'door-open' : 'door-closed'} size={16} color={statusColor} />
-                            <Text numberOfLines={1} style={{ fontSize: tokens.fontSize.sm, fontWeight: tokens.fontWeight.semibold, color: statusColor, marginLeft: 4, flexShrink: 1 }}>
-                                {statusText}
-                            </Text>
-                            
-                            {isOpen && rate !== null && (
-                                <>
-                                    <View style={{ flex: 1, height: 6, backgroundColor: theme.greyBackground, borderRadius: 3, overflow: 'hidden', marginHorizontal: tokens.space.sm }}>
-                                        <View style={{ width: `${rate}%`, height: '100%', backgroundColor: statusColor, borderRadius: 3 }} />
-                                    </View>
-                                    <Text style={{ fontSize: tokens.fontSize.xs, color: theme.fontSecondary, fontWeight: tokens.fontWeight.bold }}>
-                                        {`${rate}%`}
-                                    </Text>
-                                </>
-                            )}
-                        </View>
-                    </View>
-                </TouchableOpacity>
-            </Reanimated.View>
+            <LibrarySectionCard
+                item={item}
+                affluenceData={affluences[item.id]}
+                navigation={navigation}
+                isFavorite={favBu.includes(item.id)}
+                onToggleFavorite={toggleFavBu}
+            />
         );
     };
